kexec fix

+++ build-scripts/rootfs.sh
@@ -67,7 +67,7 @@ chmod 700 $ROOT_SSH
 #
 # Install kexec-tools
 #
-if [ "$KEXEC_TOOLS" != "" ]; then
+if [ "$KEXEC_TOOLS" != "" -a -f $OUT_DIR/$KEXEC_TOOLS.$TC_CFG._install.tar.gz ]; then
 	run_with_check tar vfx $OUT_DIR/$KEXEC_TOOLS.$TC_CFG._install.tar.gz
 	run_with_check cp -a $KEXEC_TOOLS.$TC_CFG._install/* $ROOTFS
 fi
